a2a/client/requirements.txt
a2a-sdk>=0.3.1
python-dotenv>=1.0.0
click>=8.1.0